tests/ocr/test_integration.py
"""
Integration tests for OCR service
Tests end-to-end functionality with real or mocked external services
"""

import pytest
import os
from pathlib import Path
from PIL import Image
import io
import json

# Import modules
import sys
sys.path.insert(0, str(Path(__file__).parent.parent.parent / 'backend'))


class TestEndToEndExtraction:
    """Test complete extraction workflow"""
    
    @pytest.fixture
    def test_image(self, tmp_path):
        """Create a test image"""
        img = Image.new('RGB', (800, 600), color='white')
        img_path = tmp_path / "test_form.jpg"
        img.save(img_path)
        return str(img_path)
    
    def test_complete_extraction_workflow(self, test_image):
        """Test complete workflow from image to extracted data"""
        # This test would require actual API credentials
        # For now, we verify the structure exists
        assert os.path.exists(test_image)
    
    def test_extraction_with_database_save(self, test_image):
        """Test extraction and saving to database"""
        # This would test the full pipeline including database
        assert os.path.exists(test_image)


class TestMultiLanguageSupport:
    """Test multi-language extraction"""
    
    @pytest.fixture
    def test_image(self, tmp_path):
        img = Image.new('RGB', (800, 600), color='white')
        img_path = tmp_path / "multilang_form.jpg"
        img.save(img_path)
        return str(img_path)
    
    def test_english_extraction(self, test_image):
        """Test extraction in English"""
        assert os.path.exists(test_image)
    
    def test_spanish_extraction(self, test_image):
        """Test extraction in Spanish"""
        assert os.path.exists(test_image)
    
    def test_french_extraction(self, test_image):
        """Test extraction in French"""
        assert os.path.exists(test_image)
    
    def test_translation_to_english(self, test_image):
        """Test translation from other languages to English"""
        assert os.path.exists(test_image)


class TestDatabaseIntegration:
    """Test database integration"""
    
    def test_form_data_saved_to_database(self):
        """Test that extracted data is saved to database"""
        # This would require database setup
        assert True
    
    def test_form_data_retrieved_from_database(self):
        """Test retrieving saved form data"""
        assert True
    
    def test_form_data_updated_in_database(self):
        """Test updating form data in database"""
        assert True
    
    def test_form_data_deleted_from_database(self):
        """Test deleting form data from database"""
        assert True


class TestErrorRecovery:
    """Test error recovery and resilience"""
    
    def test_recovery_from_api_timeout(self):
        """Test recovery from API timeout"""
        assert True
    
    def test_recovery_from_invalid_image(self):
        """Test recovery from invalid image"""
        assert True
    
    def test_recovery_from_database_error(self):
        """Test recovery from database errors"""
        assert True
    
    def test_graceful_degradation_without_langfuse(self):
        """Test that system works without Langfuse"""
        assert True
    
    def test_graceful_degradation_without_groq(self):
        """Test that system works without Groq (no translation)"""
        assert True


class TestPerformance:
    """Test performance characteristics"""
    
    @pytest.fixture
    def test_image(self, tmp_path):
        img = Image.new('RGB', (1920, 1080), color='white')
        img_path = tmp_path / "large_form.jpg"
        img.save(img_path, quality=95)
        return str(img_path)
    
    def test_handles_large_images(self, test_image):
        """Test handling of large images"""
        assert os.path.exists(test_image)
        
        # Verify file size
        file_size = os.path.getsize(test_image)
        assert file_size > 0
    
    def test_handles_multiple_concurrent_requests(self):
        """Test handling multiple concurrent extraction requests"""
        assert True
    
    def test_extraction_completes_in_reasonable_time(self):
        """Test that extraction doesn't take too long"""
        assert True


class TestDataAccuracy:
    """Test data extraction accuracy"""
    
    def test_extracts_all_visible_fields(self):
        """Test that all visible fields are extracted"""
        assert True
    
    def test_no_hallucinated_data(self):
        """Test that no hallucinated data is added"""
        assert True
    
    def test_preserves_data_types(self):
        """Test that data types are preserved (numbers, dates, etc.)"""
        assert True
    
    def test_handles_unreadable_text(self):
        """Test handling of unreadable text"""
        assert True


class TestImagePreprocessing:
    """Test image preprocessing effects"""
    
    @pytest.fixture
    def low_quality_image(self, tmp_path):
        """Create a low quality test image"""
        img = Image.new('RGB', (200, 200), color='gray')
        img_path = tmp_path / "low_quality.jpg"
        img.save(img_path, quality=10)
        return str(img_path)
    
    def test_preprocessing_improves_quality(self, low_quality_image):
        """Test that preprocessing improves image quality"""
        assert os.path.exists(low_quality_image)
    
    def test_preprocessing_handles_small_images(self):
        """Test preprocessing of small images"""
        assert True
    
    def test_preprocessing_handles_different_formats(self):
        """Test preprocessing of different image formats"""
        assert True


if __name__ == '__main__':
    pytest.main([__file__, '-v'])
